service-workers/service-worker/resource-timing-bodySize.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/common/utils.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>

function test_scenario(name) {
    promise_test(async (t) => {
        const uid = token();
        const worker_url = `resources/fetch-response.js?uid=${uid}`;
        const scope = `resources/fetch-response.html?uid=${uid}`;
        const iframe = document.createElement('iframe');
        iframe.src = `${scope}&path=${name}`;
        const registration = await service_worker_unregister_and_register(t, worker_url, scope);
        t.add_cleanup(() => registration.unregister());
        t.add_cleanup(() => iframe.remove());
        await wait_for_state(t, registration.installing, 'activated');
        const waitForMessage = new Promise(resolve => window.addEventListener('message', ({data}) => resolve(data)));
        document.body.appendChild(iframe);
        const {buffer, entry} = await waitForMessage;
        assert_greater_than(buffer.byteLength, 0);
        assert_equals(entry.decodedBodySize, buffer.byteLength);
        assert_equals(entry.encodedBodySize, buffer.byteLength);
    }, `Constructed response body size: ${name}`);
}

['constructed', 'forward', 'stream'].forEach(test_scenario);

</script>
